requirements.txt
openpyxl>=3.1.0
anthropic>=0.40.0
openai>=1.0.0
pydantic>=2.0
rapidfuzz>=3.0
pyyaml>=6.0
rich>=13.0
click>=8.0
flask>=3.0
eval_type_backport>=0.3.0
